requirements.txt
amqp==5.0.3
asgiref==3.3.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.3
django-allauth==0.44.0
django-bootstrap4==2.3.1
django-ckeditor==6.0.0
django-ckeditor-widget==0.1
django-filter==2.4.0
django-js-asset==1.2.2
django-rest-auth==0.9.5
django-seed==0.2.2
django-taggit==1.3.0
djangorestframework==3.12.2
Faker==5.0.1
idna==2.10
kombu==5.0.2
Markdown==3.3.3
oauthlib==3.1.0
Pillow==8.0.1
prompt-toolkit==3.0.14
pycparser==2.20
PyJWT==2.0.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
soupsieve==2.0.1
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
